+++ packages/deamon/src/utils.rs
@@ -5,6 +5,22 @@ use std::fs::File;
 use std::io;
 use std::process::{Command, Output};
 
+pub fn del_dir_if_exists(folder_path: String) -> Result<()> {
+    let path = std::path::Path::new(&folder_path);
+    if path.exists() {
+        std::fs::remove_dir_all(folder_path.clone())?;
+    }
+    Ok(())
+}
+
+pub fn create_dir_if_not_exists(folder_path: String) -> Result<()> {
+    let path = std::path::Path::new(&folder_path);
+    if !path.exists() {
+        std::fs::create_dir_all(folder_path.clone())?;
+    }
+    Ok(())
+}
+
 pub fn get_ps_snapshot() -> Result<String> {
     let ps_output = Command::new("ps").arg("aux").output()?;
 
@@ -60,6 +76,36 @@ pub fn check_and_create_file(file_path: &str) -> io::Result<File> {
         let file = File::open(file_path)?;
         return Ok(file);
     }
+}
+
+pub fn list_mount_info() -> Result<Vec<String>> {
+    let output = Command::new("mount").output()?;
+
+    let stdout = String::from_utf8_lossy(&output.stdout);
+    let mount_lines: Vec<&str> = stdout.split('\n').collect();
+
+    let filtered_mounts: Vec<String> = mount_lines
+        .iter()
+        .filter(|&&line| {
+            line.contains("node_modules")
+                && (line.starts_with("fuse-t:/") || line.starts_with("overlay"))
+        })
+        .map(|&line| {
+            let parts: Vec<&str> = line.split_whitespace().collect();
+            let device = parts[0].to_string();
+            let mount_point = parts[2].to_string();
+            // MountInfo {
+            //     device,
+            //     mount_point,
+            // }
+            mount_point
+        })
+        .collect();
+
+    let mut sorted_mounts = filtered_mounts;
+    sorted_mounts.sort_by(|a, b| a.cmp(&b));
+
+    Ok(sorted_mounts)
 }
 
 #[cfg(test)]